+++ src/runtime/input.rs
@@ -1,15 +1,15 @@
 pub trait Input {
-    fn get(&self, position: usize) -> Option<u8>;
+    fn get(&self, position: u32) -> Option<u8>;
 
-    fn len(&self) -> usize;
+    fn len(&self) -> u32;
 }
 
 impl Input for [u8] {
-    fn get(&self, position: usize) -> Option<u8> {
-        self.get(position).copied()
+    fn get(&self, position: u32) -> Option<u8> {
+        self.get(position as usize).copied()
     }
 
-    fn len(&self) -> usize {
-        self.len()
+    fn len(&self) -> u32 {
+        self.len() as u32
     }
 }
